# 2swagger_collection.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CSC309PP1-EndpointsTEST",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "noauthAuth": {
        "type": "http",
        "scheme": "noauth"
      }
    }
  },
...
  "paths": {
...
    "/api/flights/details": {
      "get": {
        "tags": ["flights"],
        "summary": "Get-flight-details",
        "security": [
          {
            "noauthAuth": []
          }
        ],
        "parameters": [
          {
            "name": "origin",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Test Case 1 (returns only both one way and round trips)",
            "example": "Toronto"
          },
          {
            "name": "destination",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Test Case 1",
            "example": "Brussels"
          },
          {
            "name": "date",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Test Case 1",
            "example": "2024-11-17"
          },
          {
            "name": "flight_type",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Test Case 1",
            "example": "round_trip"
          },
          {
            "name": "round_trip_date",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": "2024-11-17"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
...
    "/api/hotel/user/room-booking": {
      "post": {
        "tags": ["hotel > User"],
        "summary": "create-room-booking",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "creditCardInfo": "5000-00adsfa0-0000-0000",
                  "checkInDate": "2025-03-25",
                  "checkOutDate": "2025-03-29",
                  "hotelId": 1,
                  "roomType": "Family"
                }
              }
            }
          }
        },
        "security": [
          {
            "noauthAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
...
  }
}